Lots of stuff, read extended description
The main menu now works for all purposes:
The options menu is just the main menu with a green tint.
The exit button closes the game.
The New Game button gives you the same as the Continue button for now,
which is my awesome drawing of what the game might look like in a few
days.